+++ nodes/web_viewer_nodes.py
@@ -6,8 +6,13 @@ import numpy as np
 from PIL import Image
 import folder_paths
 from .image_nodes import VrchImageSaverNode
+from .audio_nodes import VrchAudioSaverNode
+from .node_utils import VrchNodeUtils
 
+# Define the category for organizational purposes
 CATEGORY = "vrch.ai/viewer"
+# The default server address
+DEFAULT_SERVER = "127.0.0.1:8188"
 
 class VrchWebViewerNode:
     @classmethod
@@ -15,7 +20,7 @@ class VrchWebViewerNode:
         return {
             "required": {
                 "mode": (["image", "flipbook", "audio", "depthmap"], {"default": "image"}),
-                "server": ("STRING", {"default": "127.0.0.1:8188", "multiline": False, "dynamicPrompts": False}),
+                "server": ("STRING", {"default": DEFAULT_SERVER, "multiline": False, "dynamicPrompts": False}),
                 "ssl": ("BOOLEAN", {"default": False}),
                 "filename": ("STRING", {"default": "web_viewer_image.jpeg", "multiline": False, "dynamicPrompts": False}),
                 "path": ("STRING", {"default": "web_viewer", "multiline": False, "dynamicPrompts": False}),
@@ -53,7 +58,7 @@ class VrchImageWebViewerNode(VrchImageSaverNode):
             "required": {
                 "images": ("IMAGE",),
                 "channel": (["1", "2", "3", "4", "5", "6", "7", "8"], {"default": "1"}),
-                "server": ("STRING", {"default": "127.0.0.1:8188", "multiline": False}),
+                "server": ("STRING", {"default": DEFAULT_SERVER, "multiline": False}),
                 "ssl": ("BOOLEAN", {"default": False}),
                 "window_width": ("INT", {"default": 1280, "min": 100, "max": 10240}),
                 "window_height": ("INT", {"default": 960, "min": 100, "max": 10240}),
@@ -113,7 +118,7 @@ class VrchImageFlipBookWebViewerNode(VrchImageSaverNode):
                 "images": ("IMAGE",),
                 "channel": (["1", "2", "3", "4", "5", "6", "7", "8"], {"default": "1"}),
                 "number_of_images": ("INT", {"default": 4, "min": 1, "max": 99}),
-                "server": ("STRING", {"default": "127.0.0.1:8188", "multiline": False}),
+                "server": ("STRING", {"default": DEFAULT_SERVER, "multiline": False}),
                 "ssl": ("BOOLEAN", {"default": False}),
                 "window_width": ("INT", {"default": 1280, "min": 100, "max": 10240}),
                 "window_height": ("INT", {"default": 960, "min": 100, "max": 10240}),
@@ -164,6 +169,62 @@ class VrchImageFlipBookWebViewerNode(VrchImageSaverNode):
             m.update(image_bytes)
         return m.hexdigest()
     
+class VrchAudioWebViewerNode(VrchAudioSaverNode):
+    
+    @classmethod
+    def INPUT_TYPES(cls):
+        return {
+            "required": {
+                "audio": ("AUDIO",),
+                "channel": (["1", "2", "3", "4", "5", "6", "7", "8"], {"default": "1"}),
+                "server": ("STRING", {"default": DEFAULT_SERVER, "multiline": False}),
+                "ssl": ("BOOLEAN", {"default": False}),
+                "window_width": ("INT", {"default": 1280, "min": 100, "max": 10240}),
+                "window_height": ("INT", {"default": 960, "min": 100, "max": 10240}),
+                "show_url": ("BOOLEAN", {"default": False}),
+                "url": ("STRING", {"default": "", "multiline": True}),
+            }
+        }
+
+    RETURN_TYPES = ("AUDIO",)
+    RETURN_NAMES = ("AUDIO",)
+    FUNCTION = "save_and_view_audio"
+    OUTPUT_NODE = True
+    CATEGORY = CATEGORY
+
+    def __init__(self):
+        # The output directory where audio will be saved
+        self.output_dir = folder_paths.output_directory
+
+    def save_and_view_audio(self, audio, channel, server, ssl, window_width, window_height, show_url, url):
+        # Save the audio into "web_viewer" directory with filename "{channel}.mp3"
+        output_path = os.path.join(self.output_dir, "web_viewer")
+        os.makedirs(output_path, exist_ok=True)
+
+        filename = f"channel_{channel}"
+        self.save_audio(
+            audio=audio,
+            filename=filename,
+            path="web_viewer",
+            extension="mp3"
+        )
+
+        return (audio,)
+
+    @classmethod
+    def IS_CHANGED(cls, audio, **kwargs):
+        m = hashlib.sha256()
+        # Convert image to bytes and update the hash
+        if isinstance(audio, torch.Tensor):
+            audio_bytes = audio.cpu().numpy().tobytes()
+        elif isinstance(audio, np.ndarray):
+            audio_bytes = audio.tobytes()
+        else:
+            # Handle other types if necessary
+            audio_bytes = bytes()
+        m.update(audio_bytes)
+        return m.hexdigest()
+    
 class VrchVideoWebViewerNode:
     @classmethod
     def INPUT_TYPES(cls):
@@ -171,7 +232,7 @@ class VrchVideoWebViewerNode:
             "required": {
                 "filename": ("STRING", {"forceInput": True}),
                 "channel": (["1", "2", "3", "4", "5", "6", "7", "8"], {"default": "1"}),
-                "server": ("STRING", {"default": "127.0.0.1:8188", "multiline": False}),
+                "server": ("STRING", {"default": DEFAULT_SERVER, "multiline": False}),
                 "ssl": ("BOOLEAN", {"default": False}),
                 "window_width": ("INT", {"default": 1280, "min": 100, "max": 10240}),
                 "window_height": ("INT", {"default": 960, "min": 100, "max": 10240}),